Add new test workflow; remove .NET workflow
The .NET workflow defined in `dotnet.yml` has been removed. This workflow was responsible for building and testing a .NET project.

A new workflow has been added in `test.yml`. This workflow is named `.test` and is triggered on pushes and pull requests to the `main` branch.

The new workflow in `test.yml` includes two jobs:
- `testCSharp`: Tests C# exercises by setting up a .NET environment, restoring dependencies, building the project, and running tests.
- `testJavaScript`: Tests JavaScript exercises using a matrix strategy for Node.js version 22.x. It sets up a Node.js environment, installs dependencies, builds the project if necessary, and runs tests in subdirectories containing `package.json` files.