Add test, watch, and lint scripts to package.json
Updated the `test` script in the `package.json` file to run Jest tests on all files in the project instead of a placeholder command. Added a new `watch` script to run Jest in watch mode, allowing tests to re-run automatically when files change. Also added a new `lint` script to run ESLint on the entire project to check for code quality and style issues.